# packages/api-zod/src/generated/types/agentInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentInfoEndpoints } from "./agentInfoEndpoints";

export interface AgentInfo {
  name?: string;
  version?: string;
  description?: string;
  endpoints?: AgentInfoEndpoints;
  capabilities?: string[];
}