# packages/api-zod/src/generated/types/agentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentFinding } from "./agentFinding";

export interface AgentResult {
  score: number;
  confidence: number;
  findings: AgentFinding[];
  summary: string;
  label: string;
  weight: number;
}