# packages/api-zod/src/generated/types/eASAttestation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EASAttestationAttestation } from "./eASAttestationAttestation";

export interface EASAttestation {
  success?: boolean;
  attestation?: EASAttestationAttestation;
  attester?: string;
  schema?: string;
  type?: string;
  note?: string | null;
}