# packages/api-zod/src/generated/types/evaluationAgents.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentResult } from "./agentResult";

export interface EvaluationAgents {
  claimVerification: AgentResult;
  evidenceQuality: AgentResult;
  impactMagnitude: AgentResult;
}