# packages/api-zod/src/generated/types/extractRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractRequest {
  text?: string | null;
  url?: string | null;
}