# packages/api-zod/src/generated/types/hypercertDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperEval API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttestationsDetail } from "./attestationsDetail";
import type { FractionsDetail } from "./fractionsDetail";
import type { HypercertContract } from "./hypercertContract";
import type { HypercertMetadata } from "./hypercertMetadata";

export interface HypercertDetail {
  hypercert_id: string;
  uri?: string | null;
  creator_address?: string | null;
  units?: string | null;
  contract?: HypercertContract | null;
  metadata?: HypercertMetadata | null;
  attestations?: AttestationsDetail | null;
  fractions?: FractionsDetail | null;
}